fix(datadragon): adjust usage of name and id when getting champion
previously the name of the champion was used to request detailed data for the champion. The DataDragon API however accepts the ID of champion instead. In many cases there is no difference between the ID and the Name. Champions with spaces or other non-alphabetic characters in their name were unable to be retrieved using the GetChampion method. This commit refactors usage of identifier for champions from name to ID.